Refactor: 포인트 관련 API에서 사용하는 타입 분리 #SCRUM-46
