+++ src/shared/api/wallets.ts
@@ -1,35 +1,5 @@
 import { baseApi } from "./base";
-
-// TransactionType Enum
-export type TransactionType = 'EARNED' | 'USED' | 'MAINTAINED' | 'EXCHANGED';
-
-// PointType Enum
-export type PointType = 'GAME' | 'STOCK' | 'EXCHANGE' | 'ATTENDANCE';
-
-// GameType Enum
-export type GameType = 'CARD_FLIP' | 'OX_QUIZ' | 'WORD_QUIZ';
-
-// Wallet Interface
-export interface Wallet {
-  memberId: number;
-  currentPoints: number;
-  currentCoins: number;
-}
-
-// ExchangeRequest Interface
-export interface ExchangeRequest {
-  pointsExchanged: number;
-}
-
-// MiniGameTransaction Interface
-export interface MiniGameTransaction {
-  memberId: number;
-  transactionType: TransactionType;
-  gameType: GameType;
-  points: number;
-  isWin: boolean;
-  createdAt: string; // ISO 8601 format
-}
+import { MiniGameTransaction, Wallet, ExchangeRequest } from "@/features/minigame/points/types/pointTypes";
 
 // Wallet API
 export const walletApi = {
